MedicalChatbotRAG/requirements.txt
langchain
langchain-core
langchain-community
langchain-text-splitters
pypdf

faiss-cpu

langchain-huggingface
transformers
huggingface-hub

python-dotenv

streamlit
